Programme to find symmetric difference
Function sym() can take multiple array arguments. First it removes duplicates from within arrays, so if one array had [1, 1, 2] it would be set to [1, 2].
The new function compare() was written to compare an array of 2 arrays. Using filter with indexOf = -1, unique values can be found. The arrays are then concatenated.
Because the 2 arrays become 1, the function only ever needs to work on 2 arrays, no matter how many arguments are supplied, as 4 arguments will become 1 as the function continues. Finally, the array is sorted.