Programme to update an inventory.
Given two arrays of nested arrays, the inventory update programme finds matching entries in the newInv and curInv arrays. If there's a match, the numbers are added together.
If no match, the non matching component of newInv is pushed onto the end of curInv.

To sort the arrays, two loops are used. One going from [0] to [-1] in the array, to bring values towards the end, and another going from [-1] to [0] to bring values towards the front.